✨ Implemented removal of attributes post-reweave
Turns out the reference list in dnlib is calculated based on actual type
references to the exported types of another assembly. If this accurately
covers how a .net module works, this also explains why things like
xaml-only project to project references are such a shit show to work with.
Makes sense though...

Resolves #4